pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jira-mcp-server"
version = "0.1.0"
description = "MCP Server for Jira Weekly Status Reporting"
authors = [
    {name = "Engineering Team", email = "team@example.com"},
]
dependencies = [
    "mcp>=1.0.0",
    "jira>=3.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "click>=8.0.0",
    "markdown>=3.5.0",
    "jinja2>=3.1.0",
    "python-dateutil>=2.8.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
jira-mcp-server = "jira_mcp_server.main:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=jira_mcp_server --cov-report=term-missing"